Move pkg/utils.go to pkg/maps/maps.go.
